move current year function to utils
