+++ intranet/apps/events/models.py
@@ -10,6 +10,7 @@ from .notifications import event_approval_request
 from ..announcements.models import Announcement
 from ..eighth.models import EighthScheduledActivity
 from ..users.models import User
+from ...utils.date import is_current_year
 
 
 class Link(models.Model):
@@ -161,12 +162,7 @@ class Event(models.Model):
     @property
     def is_this_year(self):
         """Return whether the event was created after September 1st of this school year."""
-        now = datetime.now().date()
-        ann = self.added.date()
-        if now.month < 9:
-            return ((ann.year == now.year and ann.month < 9) or (ann.year == now.year - 1 and ann.month >= 9))
-        else:
-            return (ann.year == now.year and ann.month >= 9)
+        return is_current_year(self.added.date())
 
     @property
     def dashboard_type(self):
